Split allocation-agnostic code into 'common' module.
